Browse: make /browse/$lang a lazy route
Split the route into an eager config half (browse.$lang.tsx — validateSearch,
beforeLoad, loader, staticData, exported TabSchema/BrowseSearch) and a lazy
component half (browse.$lang.lazy.tsx via createLazyFileRoute). The whole
browse UI — component plus its -browse-discover helper tiles — now ships in a
separate on-demand chunk (~6 kB gzipped) instead of the eager bundle.
